Fixed bug: com.beimin.eveapi.model.shared.Contact.labelMask :: not int > is long com.beimin.eveapi.model.shared.ContactLabel.labelID :: not int > is long

+++ src/main/java/com/beimin/eveapi/model/shared/ContactLabel.java
@@ -1,14 +1,14 @@
 package com.beimin.eveapi.model.shared;
 
 public class ContactLabel {
-    private int labelID;
+    private long labelID;
     private String name;
 
-    public int getLabelID() {
+    public long getLabelID() {
         return labelID;
     }
 
-    public void setLabelID(final int labelID) {
+    public void setLabelID(final long labelID) {
         this.labelID = labelID;
     }
